add recipient resolvers
